commit main del victor

+++ DSB/src/main/java/com/DevSalud/DSB/Service/UserServices.java
@@ -1,5 +1,8 @@
 package com.DevSalud.DSB.Service;
 
+import com.DevSalud.DSB.Exception.GeneralServiceException;
+import com.DevSalud.DSB.Exception.NoDataFoundException;
+import com.DevSalud.DSB.Exception.ValidateServiceException;
 import com.DevSalud.DSB.Model.UserModel;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
@@ -13,26 +16,78 @@ public class UserServices {
     private UserRepository userRepository;
 
     /**
-     * @return Todos los Usuarios encontrados
+     * Obtiene todos los usuarios.
+     * @return Lista de usuarios.
+     * @throws NoDataFoundException si no se encuentran usuarios.
      */
     public List<UserModel> getAllUsers() {
-        return userRepository.findAll();
+        List<UserModel> users = userRepository.findAll();
+        if (users.isEmpty()) {
+            throw new NoDataFoundException("El usuario con ID no puede estar vacia");
+        }
+        return users;
     }
 
     /**
-     * @param Model
-     * @return Guardado del Modelo User
+     * Guarda un usuario.
+     * @param user El modelo de usuario a guardar.
+     * @return El usuario guardado.
      */
-    public UserModel saveUsers(UserModel Model) {
-        return userRepository.save(Model);
+    public void saveOrUpdateUsers(UserModel user) {
+        userRepository.save(user);
     }
 
+    /**
+     * Obtiene un usuario por su ID.
+     * @param id El ID del usuario.
+     * @return El usuario encontrado.
+     * @throws ValidateServiceException si el ID del usuario es nulo.
+     * @throws NoDataFoundException si no se encuentra un usuario con el ID proporcionado.
+     */
     public UserModel getUserById(Long id) {
-        return userRepository.findById(id).orElse(null);
+        if (id == null) {
+            throw new ValidateServiceException("El usuario con ID no puede estar vacia");
+        }
+        return userRepository.findById(id).orElseThrow(() -> new NoDataFoundException("El usuaio no pudo ser encontrado con la Id: " + id));
     }
 
+    /**
+     * Elimina un usuario por su ID.
+     * @param id El ID del usuario.
+     * @throws ValidateServiceException si el ID del usuario es nulo.
+     * @throws GeneralServiceException si ocurre un error al eliminar el usuario.
+     */
     public void deleteUserById(Long id) {
-        userRepository.deleteById(id);
+        if (id == null) {
+            throw new ValidateServiceException("El usuario con ID no puede estar vacia");
+        }
+        try {
+            userRepository.deleteById(id);
+        } catch (Exception e) {
+            throw new GeneralServiceException("Error al eliminar el usuario", e.getCause());
+        }
+    }
+
+    /**
+     * Autentica un usuario.
+     *
+     * @param loginModel El modelo de inicio de sesión del usuario.
+     * @return El usuario autenticado.
+     * @throws ValidateServiceException si el nombre de usuario, correo electrónico o contraseña son inválidos.
+     */
+    public UserModel loginUser(UserModel loginModel) {
+        if ((loginModel.getUsername() == null || loginModel.getUsername().isEmpty()) &&
+                (loginModel.getEmailAddress() == null || loginModel.getEmailAddress().isEmpty())) {
+            throw new ValidateServiceException("Debe proporcionar un nombre de usuario o un correo electrónico");
+        }
+        if (loginModel.getPassword() == null || loginModel.getPassword().isEmpty()) {
+            throw new ValidateServiceException("La contraseña no puede estar vacía");
+        }
+        UserModel user = userRepository.findByUsernameOrEmailAddress(loginModel.getUsername(), loginModel.getEmailAddress());
+        if (user == null || !user.getPassword().equals(loginModel.getPassword())) {
+            throw new ValidateServiceException("Credenciales inválidas");
+        }
+        return user;
     }
 
 }
